separates locations and connection into separate vms
